refactor(react): execute 测试抽到独立文件 execute_tests.rs
- execute.rs 收敛为纯实现（3474 行 → 1104 行），测试经 #[path]
  平铺为 react/execute_tests.rs，保持子模块可见性，逻辑零改动
